+++ Options.lua
@@ -7,10 +7,11 @@ local DeathKnight = Engraved.DeathKnight
 --[[ Defaults ]]-- 
 
 Engraved.Defaults = {
-	-- AtRestAlpha = 0,
+	-- AtRestAlpha = 0, 
+	ChargeType = "ALMOST_READY", 
 	IsLocked = true, 
 	OutOfCombatAlpha = 0.2,
-	PrioritizeRunes = true, 
+	PrioritizeRunes = true,  -- Not yet implemented
 	RuneColor = { r = 1, g = 1, b = 1 }, 
 	RuneColorBlood = { r = 1, g = 0.2, b = 0 }, 
 	RuneColorFrost = { r = 0.5, g = 0.8, b = 1 }, 
@@ -34,91 +35,93 @@ Engraved.Defaults = {
 }
 
 function Engraved:SetClassDefaults(class)
-	local defaults = Engraved.Defaults;
-	if ( class == "DEATHKNIGHT" ) then
-		defaults.RuneTheme = "FROSTMOURNE"; 
-		defaults.RuneColor = { r = 0.5, g = 0.8, b = 1.0 }; 
-		defaults.RuneFrameSize = { 404, 80 }; 
-		defaults.RunePositions[1] = { "CENTER", -162, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -97, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", -32, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 32, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 97, 0 }; 
-		defaults.RunePositions[6] = { "CENTER", 162, 0 }; 
-	elseif ( class == "ROGUE" ) then
-		defaults.RuneTheme = "ICECROWN"; 
-		defaults.RuneColor = { r = 1.0, g = 0.2, b = 0.0 }; 
-		defaults.RuneFrameSize = { 340, 80 };
-		defaults.RunePositions[1] = { "CENTER", -130, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -65, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 0, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 65, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 130, 0 }; 
-		defaults.RuneSizes = { 80, 80, 80, 80, 80, 40, 40, 40 };
-	elseif ( class == "DRUID" ) then
-		defaults.RuneTheme = "NEXUS"; 
-		defaults.RuneColor = { r = 1.0, g = 0.2, b = 0.0 }; 
-		defaults.RuneFrameSize = { 340, 80 };
-		defaults.RunePositions[1] = { "CENTER", -130, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -65, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 0, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 65, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 130, 0 }; 
-	elseif ( class == "WARLOCK" ) then
-		defaults.RuneTheme = "NEXUS"; 
-		defaults.RuneColor = { r = 0.50, g = 0.32, b = 0.55 }; 
-		defaults.RuneFrameSize = { 340, 80 };
-		defaults.RunePositions[1] = { "CENTER", -130, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -65, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 0, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 65, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 130, 0 }; 
-	elseif ( class == "MAGE" ) then
-		defaults.RuneTheme = "ELDRITCH"; 
-		defaults.RuneColor = { r = 0.5, g = 0.4, b = 1 }; 
-		defaults.RuneFrameSize = { 275, 80 };
-		defaults.RunePositions[1] = { "CENTER", -97, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -32, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 32, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 97, 0 }; 
-	elseif ( class == "PALADIN" ) then
-		defaults.RuneTheme = "ULDUAR"; 
-		defaults.RuneColor = { r = 0.95, g = 0.90, b = 0.60 }; 
-		defaults.RuneFrameSize = { 340, 80 };
-		defaults.RunePositions[1] = { "CENTER", -130, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -65, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 0, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 65, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 130, 0 }; 
-	elseif ( class == "MONK" ) then
-		defaults.RuneTheme = "ULDUAR"; 
-		defaults.RuneColor = { r = 0.71, g = 1.0, b = 0.92 }; 
-		defaults.RuneFrameSize = { 404, 80 }; 
-		defaults.RunePositions[1] = { "CENTER", -162, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -97, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", -32, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 32, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 97, 0 }; 
-		defaults.RunePositions[6] = { "CENTER", 162, 0 }; 
-	elseif ( class == "EVOKER" ) then
-		defaults.RuneTheme = "NEXUS"; 
-		defaults.RuneColor = { r = 0.6, g = 0.8, b = 1.0 }; 
-		defaults.RuneFrameSize = { 340, 80 };
-		defaults.RunePositions[1] = { "CENTER", -130, 0 }; 
-		defaults.RunePositions[2] = { "CENTER", -65, 0 }; 
-		defaults.RunePositions[3] = { "CENTER", 0, 0 }; 
-		defaults.RunePositions[4] = { "CENTER", 65, 0 }; 
-		defaults.RunePositions[5] = { "CENTER", 130, 0 }; 
+	--local _, class = UnitClass("player")
+	local defaults = Engraved.Defaults
+	if class == "DEATHKNIGHT" then
+		defaults.RuneTheme = "FROSTMOURNE"
+		defaults.RuneColor = { r = 0.5, g = 0.8, b = 1.0 }
+		defaults.RuneFrameSize = { 404, 80 }
+		defaults.RunePositions[1] = { "CENTER", -162, 0 }
+		defaults.RunePositions[2] = { "CENTER", -97, 0 }
+		defaults.RunePositions[3] = { "CENTER", -32, 0 }
+		defaults.RunePositions[4] = { "CENTER", 32, 0 }
+		defaults.RunePositions[5] = { "CENTER", 97, 0 }
+		defaults.RunePositions[6] = { "CENTER", 162, 0 }
+	elseif class == "ROGUE" then
+		defaults.RuneTheme = "ICECROWN"
+		defaults.RuneColor = { r = 1.0, g = 0.2, b = 0.0 }
+		defaults.RuneFrameSize = { 340, 80 }
+		defaults.RunePositions[1] = { "CENTER", -130, 0 }
+		defaults.RunePositions[2] = { "CENTER", -65, 0 }
+		defaults.RunePositions[3] = { "CENTER", 0, 0 }
+		defaults.RunePositions[4] = { "CENTER", 65, 0 }
+		defaults.RunePositions[5] = { "CENTER", 130, 0 }
+		defaults.RuneSizes = { 80, 80, 80, 80, 80, 40, 40, 40 }
+	elseif class == "DRUID" then
+		defaults.RuneTheme = "NEXUS"
+		defaults.RuneColor = { r = 1.0, g = 0.2, b = 0.0 }
+		defaults.RuneFrameSize = { 340, 80 }
+		defaults.RunePositions[1] = { "CENTER", -130, 0 }
+		defaults.RunePositions[2] = { "CENTER", -65, 0 }
+		defaults.RunePositions[3] = { "CENTER", 0, 0 }
+		defaults.RunePositions[4] = { "CENTER", 65, 0 }
+		defaults.RunePositions[5] = { "CENTER", 130, 0 }
+	elseif class == "WARLOCK" then
+		defaults.RuneTheme = "NEXUS"
+		defaults.RuneColor = { r = 0.50, g = 0.32, b = 0.55 }
+		defaults.RuneFrameSize = { 340, 80 }
+		defaults.RunePositions[1] = { "CENTER", -130, 0 }
+		defaults.RunePositions[2] = { "CENTER", -65, 0 }
+		defaults.RunePositions[3] = { "CENTER", 0, 0 }
+		defaults.RunePositions[4] = { "CENTER", 65, 0 }
+		defaults.RunePositions[5] = { "CENTER", 130, 0 }
+	elseif class == "MAGE" then
+		defaults.RuneTheme = "ELDRITCH"
+		defaults.RuneColor = { r = 0.5, g = 0.4, b = 1 }
+		defaults.RuneFrameSize = { 275, 80 }
+		defaults.RunePositions[1] = { "CENTER", -97, 0 }
+		defaults.RunePositions[2] = { "CENTER", -32, 0 }
+		defaults.RunePositions[3] = { "CENTER", 32, 0 }
+		defaults.RunePositions[4] = { "CENTER", 97, 0 }
+	elseif class == "PALADIN" then
+		defaults.RuneTheme = "ULDUAR"
+		defaults.RuneColor = { r = 0.95, g = 0.90, b = 0.60 }
+		defaults.RuneFrameSize = { 340, 80 }
+		defaults.RunePositions[1] = { "CENTER", -130, 0 }
+		defaults.RunePositions[2] = { "CENTER", -65, 0 }
+		defaults.RunePositions[3] = { "CENTER", 0, 0 }
+		defaults.RunePositions[4] = { "CENTER", 65, 0 }
+		defaults.RunePositions[5] = { "CENTER", 130, 0 }
+	elseif class == "MONK" then
+		defaults.RuneTheme = "ULDUAR"
+		defaults.RuneColor = { r = 0.71, g = 1.0, b = 0.92 }
+		defaults.RuneFrameSize = { 404, 80 }
+		defaults.RunePositions[1] = { "CENTER", -162, 0 }
+		defaults.RunePositions[2] = { "CENTER", -97, 0 }
+		defaults.RunePositions[3] = { "CENTER", -32, 0 }
+		defaults.RunePositions[4] = { "CENTER", 32, 0 }
+		defaults.RunePositions[5] = { "CENTER", 97, 0 }
+		defaults.RunePositions[6] = { "CENTER", 162, 0 }
+	elseif class == "EVOKER" then
+		defaults.RuneTheme = "NEXUS"
+		defaults.RuneColor = { r = 0.6, g = 0.8, b = 1.0 }
+		defaults.RuneFrameSize = { 340, 80 }
+		defaults.RunePositions[1] = { "CENTER", -130, 0 }
+		defaults.RunePositions[2] = { "CENTER", -65, 0 }
+		defaults.RunePositions[3] = { "CENTER", 0, 0 }
+		defaults.RunePositions[4] = { "CENTER", 65, 0 }
+		defaults.RunePositions[5] = { "CENTER", 130, 0 }
 	end
-	Engraved.OptionsPanel.defaults = defaults;
+	Engraved.OptionsPanel.defaults = defaults
 end
 
 
---[[ Interface options panel ]]-- 
+--[[ OptionsPanel ]]-- 
 
-Engraved.OptionsPanel = Engraved:CreateOptionsPanel();
-local optionsPanel = Engraved.OptionsPanel;
-optionsPanel.savedVariablesName = "EngravedOptions";
+Engraved.OptionsPanel = Engraved:CreateOptionsPanel()
+local optionsPanel = Engraved.OptionsPanel
+
+optionsPanel.savedVariablesName = "EngravedOptions"
 -- optionsPanel.defaults set in SetClassDefaults()
 optionsPanel.defaultsFunc = function()
 	local runeFrame, options = EngravedRuneFrame, EngravedOptions;
@@ -129,9 +132,9 @@ optionsPanel.defaultsFunc = function()
 	runeFrame:SetLocked(options.IsLocked);
 end;
 
+-- Rune theme
 optionsPanel.runeTheme = optionsPanel:CreateDropDownMenu("RuneTheme");
 local runeTheme = optionsPanel.runeTheme;
--- runeTheme:SetPoint("TOPLEFT", optionsPanel.runeColor, "BOTTOMLEFT", -16, -42);
 runeTheme:SetPoint("TOPLEFT", optionsPanel.subtext, "BOTTOMLEFT", -14, -42);
 runeTheme.onValueChanged = function(value) 
 	EngravedRuneFrame:SetRuneTheme(value);
@@ -144,14 +147,15 @@ runeTheme.optionList = {
 	{text = "Eldritch", value = "ELDRITCH"} 
 }
 
+-- Rune color
 optionsPanel.runeColor = optionsPanel:CreateColorPicker("RuneColor");
 local runeColor = optionsPanel.runeColor;
--- runeColor:SetPoint("TOPLEFT", optionsPanel.subtext, "BOTTOMLEFT", 2, -34);
 runeColor:SetPoint("TOPLEFT", optionsPanel.runeTheme, "BOTTOMLEFT", 18, -28);
 runeColor.onValueChanged = function(value) 
 	EngravedRuneFrame:SetRuneColor(value);
 end
 
+-- Blood rune color
 optionsPanel.runeColorBlood = optionsPanel:CreateColorPicker("RuneColorBlood");
 local runeColorBlood = optionsPanel.runeColorBlood;
 runeColorBlood:SetPoint("TOPLEFT", optionsPanel.runeTheme, "BOTTOMLEFT", 18, -28);
@@ -160,6 +164,7 @@ runeColorBlood.onValueChanged = function(value)
 end
 runeColorBlood:Hide()
 
+-- Frost rune color
 optionsPanel.runeColorFrost = optionsPanel:CreateColorPicker("RuneColorFrost");
 local runeColorFrost = optionsPanel.runeColorFrost;
 runeColorFrost:SetPoint("LEFT", optionsPanel.runeColorBlood, "RIGHT", 128, 0);
@@ -168,6 +173,7 @@ runeColorFrost.onValueChanged = function(value)
 end
 runeColorFrost:Hide()
 
+-- Unholy rune color
 optionsPanel.runeColorUnholy = optionsPanel:CreateColorPicker("RuneColorUnholy");
 local runeColorUnholy = optionsPanel.runeColorUnholy;
 runeColorUnholy:SetPoint("LEFT", optionsPanel.runeColorFrost, "RIGHT", 128, 0);
@@ -176,6 +182,7 @@ runeColorUnholy.onValueChanged = function(value)
 end
 runeColorUnholy:Hide()
 
+-- Death rune color
 optionsPanel.runeColorDeath = optionsPanel:CreateColorPicker("RuneColorDeath");
 local runeColorDeath = optionsPanel.runeColorDeath;
 runeColorDeath:SetPoint("LEFT", optionsPanel.runeColorUnholy, "RIGHT", 128, 0);
@@ -184,6 +191,21 @@ runeColorDeath.onValueChanged = function(value)
 end
 runeColorDeath:Hide()
 
+-- Charge type
+optionsPanel.chargeType = optionsPanel:CreateDropDownMenu("ChargeType")
+local chargeType = optionsPanel.chargeType
+chargeType:SetPoint("TOPLEFT", optionsPanel.runeColorBlood, "BOTTOMLEFT", -14, -42)
+UIDropDownMenu_SetWidth(chargeType, 220)
+chargeType.onValueChanged = function(value) 
+	RuneFrame:SetChargeType(value)
+end
+chargeType.optionList = {
+	{text = "Slowly increase glow", value = "SLOW_GLOW"},
+	{text = "Glow when almost ready", value = "ALMOST_READY"}
+}
+chargeType:Hide()
+
+-- Out of combat alpha
 optionsPanel.outOfCombatAlpha = optionsPanel:CreateSlider("OutOfCombatAlpha");
 local outOfCombatAlpha = optionsPanel.outOfCombatAlpha;
 outOfCombatAlpha:SetPoint("TOPLEFT", optionsPanel.subtext, "BOTTOMLEFT", 300, -48);
@@ -195,25 +217,39 @@ outOfCombatAlpha.onValueChanged = function(value)
 	EngravedRuneFrame:UpdateShown();
 end
 
-local configButton = CreateFrame("Button", optionsPanel:GetName().."ConfigButton", optionsPanel, "UIPanelButtonTemplate");
-optionsPanel.configButton = configButton;
-configButton:SetSize(128, 22);
-configButton:SetPoint("BOTTOMLEFT", 24, 24);
-configButton.Text = _G[configButton:GetName().."Text"];
--- configButton.Text:SetText("Config text");
+-- Config mode button
+local configButton = CreateFrame("Button", optionsPanel:GetName().."ConfigButton", optionsPanel, "UIPanelButtonTemplate")
+optionsPanel.configButton = configButton
+configButton:SetSize(128, 22)
+configButton:SetPoint("BOTTOMRIGHT", optionsPanel, "BOTTOM", -12, 24)
+configButton.Text = _G[configButton:GetName().."Text"]
 configButton:SetScript("OnClick", function () 
-	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF);
-	EngravedRuneFrame:SetLocked(false);
+	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_OFF)
+	EngravedRuneFrame:SetLocked(false)
 end)
 
-local playButton = CreateFrame("Button", optionsPanel:GetName().."PlayButton", optionsPanel, "UIPanelButtonTemplate");
-optionsPanel.playButton = playButton;
-playButton:SetSize(128, 22);
-playButton:SetPoint("BOTTOMRIGHT", -24, 24);
-playButton.Text = _G[playButton:GetName().."Text"];
--- playButton.Text:SetText("Play text");
+-- Play mode button
+local playButton = CreateFrame("Button", optionsPanel:GetName().."PlayButton", optionsPanel, "UIPanelButtonTemplate")
+optionsPanel.playButton = playButton
+playButton:SetSize(128, 22)
+playButton:SetPoint("BOTTOMLEFT", optionsPanel, "BOTTOM", 12, 24)
+playButton.Text = _G[playButton:GetName().."Text"]
 playButton:SetScript("OnClick", function () 
-	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON);
-	EngravedRuneFrame:SetLocked(true);
+	PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
+	EngravedRuneFrame:SetLocked(true)
 end)
+
+local OptionsPanel = Engraved.OptionsPanel
+
+function OptionsPanel:ShowDeathKnightOptions()
+	self.runeColor:Hide()
+	self.runeColorBlood:Show()
+	self.runeColorFrost:Show()
+	self.runeColorUnholy:Show()
+	self.runeColorDeath:Show()
+	self.chargeType:Show()
+end
+
+
+